+++ tests/cases/reject_cls_super_outside.ts
@@ -1,9 +1,0 @@
-class Base {
-  ping(): number {
-    return 1;
-  }
-}
-
-function pong(): number {
-  return super.ping();
-}